+++ Scroller/useThemeScroller.js
@@ -24,9 +24,13 @@ const getFocusableBodyProps = (scrollContainerRef, contentId) => {
 
 	const setNavigableFilter = ({filterTarget}) => {
 		if (spotlightId && filterTarget) {
-			const targetClassName = (filterTarget === 'body') ? css.focusableBody : scrollbarTrackCss.thumb;
+			const isBodyFilter = (filterTarget === 'body');
+			const targetClassName = isBodyFilter ? css.focusableBody : scrollbarTrackCss.thumb;
+			// Focus should not leave scrollbar with direction key
+			const restrictOption = isBodyFilter ? 'self-only' : 'self-first';
 			Spotlight.set(spotlightId, {
-				navigableFilter: (elem) => (typeof elem === 'string' || !elem.classList.contains(targetClassName))
+				navigableFilter: (elem) => (typeof elem === 'string' || !elem.classList.contains(targetClassName)),
+				restrict: restrictOption
 			});
 			return true;
 		}